requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==4.2.20
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
psycopg2==2.9.10
python-decouple==3.8
sqlparse==0.5.3
typing-extensions==4.12.2
tzdata==2025.1
